# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ServiceHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [users]
      summary: Get notifications for current session user
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
...
  /messages:
    get:
      operationId: listMessages
      tags: [users]
      summary: Get chat messages between user and provider
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: providerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [users]
      summary: Send a chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, phone, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [user, provider, admin]
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [name, email, phone, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [user, provider]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    LoginResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
...
    Notification:
      type: object
      required: [id, userId, message, type, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        message:
          type: string
        type:
          type: string
          enum: [booking_confirmed, booking_request, review_received, booking_completed]
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Message:
      type: object
      required: [id, senderId, receiverId, content, createdAt]
      properties:
        id:
          type: integer
        senderId:
          type: integer
        receiverId:
          type: integer
        senderName:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [senderId, receiverId, content]
      properties:
        senderId:
          type: integer
        receiverId:
          type: integer
        content:
          type: string