# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ServiceHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  senderId: number;
  receiverId: number;
  content: string;
}